Change length type from int to short in setBases

+++ src/main/java/es/udc/gac/bigdec/sequence/Sequence.java
@@ -39,7 +39,7 @@ public class Sequence {
 
 	public void setBases(byte[] bases) {
 		this.bases = bases;
-		this.length = bases.length;
+		this.length = (short) bases.length;
 	}
 
 	public void setQuals(byte[] quals) {
